Update bdtMaker.py to take multiple file inputs (#69)
bdtMaker.py can now take absolute paths to directories as inputs for background and signal paths, in addition to permitting single file inputs. In the case of directories the script will add all ROOT files in the directory to the TChain. In the case of single files, the individual files will be added to the respective TChains instead.