Update bdtMaker.py with bug fix, feature names, optional SHAP analysis, and package imports (#72)
* Update bdtMaker.py to take multiple file inputs

bdtMaker.py can now take absolute paths to directories as inputs for background and signal paths, in addition to permitting single file inputs. In the case of directories the script will add all ROOT files in the directory to the TChain. In the case of single files, the individual files will be added to the respective TChains instead.

* Bug fix in bdtMaker.py

The script would print out the 'failed trigger' message every time a signal event was loaded instead of every time a signal event failed the trigger and was passed over. Now the debug message is printed out only when the event has failed the trigger.

* bdtMaker.py updates: feature_names, shap plots, package import updates

Incorporated feature names for BDT training so that the feature importance and SHAP plots display what the features are.

Included new optional functions to export training events to CSV files for further SHAP analysis, and to plot the SHAP values summary bar chart describing model performance on the test data.

Eliminated unnecessary package imports for faster runtime.

* Update bdtMaker.py for nondescript overlay pass name

* Change image format EcalVeto-3.0/bdtMaker.py

Co-authored-by: Tamas Vami <tamas.almos.vami@cern.ch>

---------

Co-authored-by: Tamas Vami <tamas.almos.vami@cern.ch>